Move all validation to input-number
